\name{lpm4i-package}
\alias{lpm4i}
\docType{package}
\title{
\packageTitle{lpm4i}
}
\description{
\packageDescription{lpm4i}
}
\author{
Riccardo Rastelli

Mantainer: Riccardo Rastelli <riccardoras@gmail.com>
}
\references{
Tafakori et al. (2020) "Measuring systemic risk and contagion in the European financial network", \url{https://arxiv.org/abs/1911.11488}
}
\keyword{ package }
